# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bihar Cyber Café Digital Service Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [Services]
      summary: List services, optionally filtered by category
      parameters:
        - name: categorySlug
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [Services]
      summary: Create a service (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /services/{id}:
    get:
      operationId: getService
      tags: [Services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateService
      tags: [Services]
      summary: Update a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteService
      tags: [Services]
      summary: Delete a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [Notifications]
      summary: List current user's notifications
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [Faqs]
      summary: List FAQs
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [Faqs]
      summary: Create an FAQ (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [Faqs]
      summary: Update an FAQ (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteFaq
      tags: [Faqs]
      summary: Delete an FAQ (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FormField:
      type: object
      required: [name, label, type, required]
      properties:
        name:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, date, select, textarea, checkbox]
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
    Service:
      type: object
      required:
        [
          id,
          categoryId,
          categorySlug,
          categoryName,
          name,
          slug,
          description,
          price,
          estimatedDays,
          formSchema,
          requiredDocuments,
          isActive,
          createdAt,
        ]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categorySlug:
          type: string
        categoryName:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        estimatedDays:
          type: integer
        formSchema:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
        requiredDocuments:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required:
        [
          categoryId,
          name,
          description,
          price,
          estimatedDays,
          formSchema,
          requiredDocuments,
        ]
      properties:
        categoryId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        estimatedDays:
          type: integer
          minimum: 0
        formSchema:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
        requiredDocuments:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        estimatedDays:
          type: integer
          minimum: 0
        formSchema:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
        requiredDocuments:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    Notification:
      type: object
      required: [id, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Faq:
      type: object
      required: [id, question, answer, category]
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        category:
          type: ["string", "null"]
    FaqInput:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
          minLength: 1
        answer:
          type: string
          minLength: 1
        category:
          type: string
    FaqUpdate:
      type: object
      properties:
        question:
          type: string
          minLength: 1
        answer:
          type: string
          minLength: 1
        category:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"