# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Cyber Café Digital Service Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}