# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Bihar Cyber Café Digital Service Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { FormField } from './formField';

export interface Service {
  id: number;
  categoryId: number;
  categorySlug: string;
  categoryName: string;
  name: string;
  slug: string;
  description: string;
  price: number;
  estimatedDays: number;
  formSchema: FormField[];
  requiredDocuments: string[];
  isActive: boolean;
  createdAt: Date;
}